got billings working, yay ^_^

+++ app/controllers/orders_controller.rb
@@ -44,20 +44,31 @@ class OrdersController < ApplicationController
   end
 
   def create
-    @cart_items = CartItem.all.order(id: :desc) # temp - CartItem.session_id(session[:id])
-    @user_id = 3 # temp -session[:user_id] if session[:user_id]
-    @billing_id = 1 # temp - params[:billing_id]
-    @order_number = order_number
-    @order = Order.new(status: "pending", confirmation_date: Time.now, order_number: @order_number, billing_id: @billing_id, user_id: @user_id)
-    if @order.save
-      @cart_items.each do |item|
-        @order_item = OrderItem.new(quantity: item.quantity, name: item.product.name, price: item.product.price*item.quantity, status: "pending", order_id: @order.id, product_id: item.product.id)
-        @order_item.save
-        item.destroy
-        @order_item.product.update(inventory: @order_item.product.inventory - @order_item.quantity)
+    @billing = Billing.new(billing_params[:billing])
+    if @billing.save
+      if current_user
+        @cart_items = CartItem.all.where(user_id: session[:user_id])
+      else
+        @cart_items = CartItem.all.where(session_id: session[:session_id])
       end
-      redirect_to order_path(@order.id)
+      @user_id = session[:user_id] if session[:user_id]
+      @billing_id = @billing.id
+      @order_number = order_number
+      @order = Order.new(status: "pending", confirmation_date: Time.now, order_number: @order_number, billing_id: @billing_id, user_id: @user_id)
+      if @order.save
+        @cart_items.each do |item|
+          @order_item = OrderItem.new(quantity: item.quantity, name: item.product.name, price: item.product.price*item.quantity, status: "pending", order_id: @order.id, product_id: item.product.id)
+          @order_item.save
+          item.destroy
+          @order_item.product.update(inventory: @order_item.product.inventory - @order_item.quantity)
+        end
+        redirect_to order_path(@order.id)
+      else
+        raise
+        render :new
+      end
     else
+      raise
       render :new
     end
   end
@@ -82,5 +93,11 @@ class OrdersController < ApplicationController
 
   def update_item_params
     params.permit(OrderItem: [:shipped])
+  end
+
+
+  def billing_params
+    params.permit(billing: [:first_name, :last_name, :cc, :cvv, :email, :billing_zip, :address, :city, :state, :zip, :user_id])
+
   end
 end
